run docker tests conditionally
